requirements.txt
alembic==1.5.8
attrs==20.3.0
cffi==1.14.5
click==7.1.2
ecdsa==0.14.1
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
hypothesis==6.8.4
Mako==1.1.4
MarkupSafe==1.1.1
pi==0.1.2
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7.2
six==1.15.0
sortedcontainers==2.3.0
SQLAlchemy==1.4.3
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
